remove fastxp and refactor fastuse
